% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{convert_to_time_series}
\alias{convert_to_time_series}
\title{Convert Data to Time Series}
\usage{
convert_to_time_series(data, s = "monthly")
}
\arguments{
\item{data}{A numeric vector or univariate time series.}

\item{s}{The frequency of the time series. Can be one of "weekly", "monthly", "quarterly", "yearly" or 52, 12, 4, 1.}
}
\value{
A time series object.
}
\description{
This function converts a numeric data vector or univariate time series to a time series object with a specified frequency.
}
\examples{
data <- c(1:48)
ts_data <- convert_to_time_series(data, s = "monthly")
}
