% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_timeseries}
\alias{generate_timeseries}
\title{Generate Synthetic Datasets}
\usage{
generate_timeseries(trend, seasonality = TRUE, noise = TRUE)
}
\arguments{
\item{trend}{The trend type for the synthetic data. Options are "linear", "quadratic", or "exponential".}

\item{seasonality}{Logical, indicating whether to include seasonality in the synthetic data.}

\item{noise}{Logical, indicating whether to include noise in the synthetic data.}
}
\value{
A list containing multiple synthetic datasets.
}
\description{
This function generates synthetic datasets with various trends and seasonality for testing purposes.
}
