% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{identify_trend}
\alias{identify_trend}
\title{Identify Trend in Data}
\usage{
identify_trend(data, s = "monthly")
}
\arguments{
\item{data}{A numeric vector or univariate time series.}

\item{s}{The frequency of the time series. Can be one of "weekly", "monthly", "quarterly", "yearly" or 52, 12, 4, 1.}
}
\value{
A list containing the identified trend and the time series object.
}
\description{
This function identifies the trend in a given numeric data vector or univariate time series.
}
\examples{
data <- c(1, 2, 3, 4, 5)
trend_result <- identify_trend(data, s = "monthly")
}
