% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalityTest.R
\name{interactive_seasonality_test}
\alias{interactive_seasonality_test}
\title{Interactive Seasonality Test}
\usage{
interactive_seasonality_test(
  data,
  s = 12,
  confidence_level = 0.05,
  summary_data = TRUE
)
}
\arguments{
\item{data}{The input data for the seasonality test.}

\item{s}{The seasonality parameter (default = 12).}

\item{confidence_level}{The desired confidence level for the statistical tests (default = 0.05).}

\item{summary_data}{Flag indicating whether to include the results summary (default = TRUE).}
}
\value{
The result of the seasonality test.
}
\description{
This function performs an interactive seasonality test on a given time series data.
}
\examples{
data <- c(1:48)
result <- interactive_seasonality_test(data, s = 12, confidence_level = 0.05, summary_data = TRUE)
print(result)
}
