% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Linear Seasonality Test}
\usage{
linear(data, s = frequency(data))
}
\arguments{
\item{data}{A numeric vector or time series object.}

\item{s}{The frequency of the time series.}
}
\value{
A list containing the Ui, Vi, and di values.
}
\description{
This function performs a seasonality test assuming a linear trend in the data.
}
\examples{
data <- c(1:48)
linear_result <- linear(data, s = 12)
}
