% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic Seasonality Test}
\usage{
quadratic(data, s)
}
\arguments{
\item{data}{A numeric vector or time series object.}

\item{s}{The frequency of the time series.}
}
\value{
A list containing the Ui, Vi, and Di values.
}
\description{
This function performs a seasonality test assuming a quadratic trend in the data.
}
\examples{
data <- ts(c(1:120), frequency = 12)
quadratic_result <- quadratic(data, s = 12)
}
