% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalityTest.R
\name{seasonality_test}
\alias{seasonality_test}
\title{Perform Seasonality Test}
\usage{
seasonality_test(
  data,
  trend,
  window = NULL,
  s = NULL,
  confidence_level = 0.05,
  summary_data = NULL
)
}
\arguments{
\item{data}{The input data for the seasonality test.}

\item{trend}{The trend type for the seasonality test. Supported options: linear, quadratic, exponential, linearMA, linearEXP.}

\item{window}{The window size for the linear moving averages trend (only applicable when trend = "linearMA").}

\item{s}{The smoothing parameter for exponential trend (only applicable when trend = "quadratic", "exponential", or "linearEXP").}

\item{confidence_level}{The desired confidence level for the statistical tests (default = 0.05).}

\item{summary_data}{Flag indicating whether to include the results summary (default = NULL).}
}
\value{
If summary_data is NULL, returns a character string indicating the presence or absence of statistically significant seasonality in the data.
If summary_data is provided, returns a list with the following elements:
- is_seasonal: A logical value indicating the presence or absence of statistically significant seasonality in the data.
- summary_data: A data frame containing the test names and p-values.
- t_summary: Summary statistics for the Student t-Distribution test.
- sign_summary: Summary statistics for the Sign Test.
- wilcox_summary: Summary statistics for the Wilcoxon Signed-Ranks Test.
}
\description{
This function performs seasonality tests on the given data based on the specified trend.
}
\examples{
# Example 1: Linear Trend with default Confidence Level and Summary Data
data <- c(10, 15, 20, 15, 10)
result <- seasonality_test(data, trend = "linear", summary_data = TRUE)
print(result)

# Example 2: Quadratic Trend with Different Confidence Level and No Summary Data
data <- c(10, 15, 20, 15, 10)
result <- seasonality_test(data, trend = "quadratic", s = 2, confidence_level = 0.01)
print(result)

# Example 3: Exponential Trend with Default Confidence Level and Summary Data
data <- c(10, 15, 20, 15, 10)
result <- seasonality_test(data, trend = "exponential", s = 0.5, summary_data = TRUE)
print(result)

# Example 4: Linear Moving Averages Trend with Different Confidence Level and Summary Data
data <- c(10, 15, 20, 15, 10)
result <- seasonality_test(data, trend = "linearMA", window = 3, s = 1, confidence_level = 0.001, summary_data = TRUE)
print(result)

# Example 5: Linear Exponential Smoothing Trend with Default Confidence Level and Summary Data
data <- c(10, 15, 20, 15, 10)
result <- seasonality_test(data, trend = "linearEXP", s = 0.1, summary_data = TRUE)
print(result)

}
