% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{exponential}
\alias{exponential}
\title{Exponential Seasonality Test}
\usage{
exponential(data, s = frequency(data))
}
\arguments{
\item{data}{A numeric vector or time series object.}

\item{s}{The frequency of the time series.}
}
\value{
A list containing the Ui, Vi, and Di values.
}
\description{
This function performs a seasonality test assuming an exponential trend in the data.
}
\examples{
data <- c(1, 2, 3, 4, 5)
exponential_result <- exponential(data, s = 12)
}
